Remove Step2Course, old MakeRooms, rename NewMakeRooms MakeRooms
