+++ client/src/Containers/Create/MakeRooms/index.js
@@ -1,5 +1,4 @@
 export { default as Step1 } from './Step1';
-export { default as Step2Course } from './Step2Course';
 export { default as Step2 } from './Step2';
 export { default as ParticipantList } from './ParticipantList';
 export { default as AssignmentMatrix } from './AssignmentMatrix';
